Finished StopDialogue.

+++ ServiceLayer/Manager/DialogsManager.cs
@@ -1,4 +1,5 @@
-﻿using MentalaisGidsAPI.Domain;
+﻿using DomainLayer.Enum;
+using MentalaisGidsAPI.Domain;
 using Microsoft.EntityFrameworkCore;
 using ServiceLayer.Interface;
 using System;
@@ -9,30 +10,36 @@ using System.Threading.Tasks;
 
 namespace ServiceLayer.Manager
 {
-    public class DialogsManager : BaseManager<Dialogs>, IDialaogsManager
+    public class DialogsManager : BaseManager<Dialogs>, IDialogsManager
     {
         private readonly MentalaisGidsContext _context;
+        private readonly IUserService _userService;
 
-        public DialogsManager(MentalaisGidsContext context) : base(context)
+        public DialogsManager(MentalaisGidsContext context, IUserService userService) : base(context)
         {
             _context = context;
+            _userService = userService;
         }
+
 
         public async Task<bool> StopDialogue(int dialogueId)
         {
             var dialogue = await _context.Dialogs.FirstOrDefaultAsync(x => x.DialogsID == dialogueId);
-            List<Zina> zinas = await _context.Zina.Where(x => x.DialogsID == dialogueId).ToListAsync();
-
             if (dialogue == null)
             {
                 return false;
             }
 
-            foreach (var zina in zinas)
+            var userId = _userService.GetUserId();
+            bool userIsInDialogue = await _context.Dialogs.AnyAsync(x => x.DialogsID == dialogueId && x.LietotajsID == userId);
+            if (!userIsInDialogue)
             {
-                _context.Zina.Remove(zina);
+                return false;
             }
 
+            List<Zina> zinas = await _context.Zina.Where(x => x.DialogsID == dialogueId).ToListAsync();
+
+            _context.Zina.RemoveRange(zinas);
             _context.Dialogs.Remove(dialogue);
             await _context.SaveChangesAsync();
